Saving progress
I think I have the code all working, it's now just updating tests and stuff. I updated the docs but they probably need to be cleaned up more.
