+++ packages/hydrogen/src/components/AddToCartButton/tests/AddToCart.test.tsx
@@ -3,7 +3,7 @@ import {CartProvider} from '../../CartProvider';
 import {mountWithProviders} from '../../../utilities/tests/shopifyMount';
 import {mountWithCartProvider} from '../../CartProvider/tests/utilities';
 
-import {ProductProvider} from '../../ProductProvider';
+import {ProductOptionsProvider} from '../../ProductOptionsProvider';
 import {AddToCartButton} from '../AddToCartButton.client';
 import {getProduct, getVariant} from '../../../utilities/tests/product';
 
@@ -79,7 +79,7 @@ describe('AddToCartButton', () => {
     });
   });
 
-  describe('when inside a ProductProvider', () => {
+  describe('when inside a ProductOptionsProvider', () => {
     describe('and an initialVariantId is present', () => {
       it('calls linesAdd with the initialVariantId', () => {
         const mockLinesAdd = jest.fn();
@@ -87,10 +87,12 @@ describe('AddToCartButton', () => {
         const selectedVariant = product?.variants?.edges?.[0]?.node;
 
         const component = mountWithCartProvider(
-          // @ts-expect-error SellingPlanAllocations isn't mocked out correctly
-          <ProductProvider data={product} initialVariantId={selectedVariant.id}>
+          <ProductOptionsProvider
+            data={product}
+            initialVariantId={selectedVariant.id}
+          >
             <AddToCartButton>Add to cart</AddToCartButton>
-          </ProductProvider>,
+          </ProductOptionsProvider>,
           {linesAdd: mockLinesAdd, cart: {id: '456'}}
         );
 
@@ -122,10 +124,9 @@ describe('AddToCartButton', () => {
         });
 
         const component = mountWithCartProvider(
-          // @ts-expect-error The mock doesn't match perfectly, fix at some point
-          <ProductProvider data={product}>
+          <ProductOptionsProvider data={product}>
             <AddToCartButton>Add to cart</AddToCartButton>
-          </ProductProvider>,
+          </ProductOptionsProvider>,
           {linesAdd: mockLinesAdd, cart: {id: '456'}}
         );
 
@@ -146,10 +147,9 @@ describe('AddToCartButton', () => {
         const product = getProduct();
 
         const component = mountWithCartProvider(
-          // @ts-expect-error The mock doesn't match perfectly, fix at some point
-          <ProductProvider data={product} initialVariantId={null}>
+          <ProductOptionsProvider data={product} initialVariantId={null}>
             <AddToCartButton>Add to cart</AddToCartButton>
-          </ProductProvider>,
+          </ProductOptionsProvider>,
           {linesAdd: mockLinesAdd}
         );
