+++ packages/hydrogen/src/components/ProductOptionsProvider/ProductOptionsProvider.client.tsx
@@ -7,14 +7,20 @@ import type {
   ProductVariant as ProductVariantType,
 } from '../../storefront-api-types';
 import type {PartialDeep} from 'type-fest';
-import {flattenConnection} from '../../utilities';
-import {getSelectedVariant} from '../../hooks/useProductOptions/helpers';
+import {
+  getSelectedVariant,
+  getOptions,
+} from '../../hooks/useProductOptions/helpers';
+import {
+  SelectedOptions,
+  ProductOptionsHookValue,
+} from '../../hooks/useProductOptions/types';
 
 type InitialVariantId = ProductVariantType['id'] | null;
 
 interface ProductOptionsProviderProps {
   /** A [Product object](https://shopify.dev/api/storefront/reference/products/product). */
-  product: PartialDeep<Product>;
+  data: PartialDeep<Product>;
   /** A `ReactNode` element. */
   children: React.ReactNode;
   /**
@@ -30,15 +36,23 @@ interface ProductOptionsProviderProps {
 
 export function ProductOptionsProvider({
   children,
-  product,
+  data: product,
   initialVariantId: explicitVariantId,
 }: ProductOptionsProviderProps) {
   // The flattened variants
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const variants = product.variants?.nodes ?? [];
+  const variants = useMemo(
+    () => product.variants?.nodes,
+    [product.variants?.nodes]
+  );
+
+  if (!isProductVariantArray(variants)) {
+    throw new Error(
+      `<ProductOptionsProvider/> requires product.variants.nodes`
+    );
+  }
 
   // All the options available for a product, based on all the variants
-  const options = product.options ?? [];
+  const options = useMemo(() => getOptions(variants), [variants]);
 
   /**
    * Track the selectedVariant within the hook. If `initialVariantId`
@@ -46,12 +60,20 @@ export function ProductOptionsProvider({
    */
   const [selectedVariant, setSelectedVariant] = useState<
     PartialDeep<ProductVariantType> | undefined | null
-  >(() => {
-    // this code block only executes the very first time 'useState' is used.
-    return getVariantBasedOnIdProp(explicitVariantId, variants);
-  });
+  >(() => getVariantBasedOnIdProp(explicitVariantId, variants));
 
-  const selectedOptions = selectedVariant?.selectedOptions ?? [];
+  /**
+   * Track the selectedOptions within the hook. If a `initialVariantId`
+   * is passed, use that to select initial options.
+   */
+  const [selectedOptions, setSelectedOptions] = useState(
+    selectedVariant?.selectedOptions
+      ? selectedVariant.selectedOptions.reduce((memo, optionSet) => {
+          memo[optionSet?.name ?? ''] = optionSet?.value ?? '';
+          return memo;
+        }, {} as SelectedOptions)
+      : {}
+  );
 
   /**
    * When the initialVariantId changes, we need to make sure we
@@ -60,8 +82,9 @@ export function ProductOptionsProvider({
    * values.
    */
   useEffect(() => {
-    setSelectedVariant(getVariantBasedOnIdProp(explicitVariantId, variants));
-  }, [explicitVariantId, variants]);
+    const variant = getSelectedVariant(variants, selectedOptions);
+    setSelectedVariant(getVariantBasedOnIdProp(variant?.id, variants));
+  }, [selectedOptions, variants]);
 
   /**
    * Allow the developer to select an option.
@@ -90,17 +113,11 @@ export function ProductOptionsProvider({
 
   const sellingPlanGroups = useMemo(
     () =>
-      sellingPlanGroupsConnection
-        ? flattenConnection(sellingPlanGroupsConnection).map(
-            (sellingPlanGroup) => ({
-              ...sellingPlanGroup,
-              sellingPlans: sellingPlanGroup?.sellingPlans
-                ? flattenConnection(sellingPlanGroup.sellingPlans)
-                : [],
-            })
-          )
-        : [],
-    [sellingPlanGroupsConnection]
+      product.sellingPlanGroups?.nodes?.map((sellingPlanGroup) => ({
+        ...sellingPlanGroup,
+        sellingPlans: sellingPlanGroup?.sellingPlans?.nodes ?? [],
+      })),
+    [product.sellingPlanGroups?.nodes]
   );
 
   /**
@@ -113,29 +130,27 @@ export function ProductOptionsProvider({
   >(undefined);
 
   const selectedSellingPlanAllocation = useMemo<
-    SellingPlanAllocation | undefined
-    // @ts-ignore The types here are broken on main, need to come back and fix them sometime
+    PartialDeep<SellingPlanAllocation> | undefined
   >(() => {
     if (!selectedVariant || !selectedSellingPlan) {
       return;
     }
 
-    if (!selectedVariant.sellingPlanAllocations) {
+    if (!selectedVariant.sellingPlanAllocations?.nodes) {
       throw new Error(
-        `You must include sellingPlanAllocations in your variants in order to calculate selectedSellingPlanAllocation`
+        `<ProductOptionsProvider/>: You must include 'sellingPlanAllocations.nodes' in your variants in order to calculate selectedSellingPlanAllocation`
       );
     }
 
-    return flattenConnection(selectedVariant.sellingPlanAllocations).find(
-      // @ts-ignore The types here are broken on main, need to come back and fix them sometime
-      (allocation) => allocation.sellingPlan.id === selectedSellingPlan.id
+    return selectedVariant.sellingPlanAllocations.nodes?.find(
+      (allocation) => allocation?.sellingPlan?.id === selectedSellingPlan.id
     );
   }, [selectedVariant, selectedSellingPlan]);
 
-  const productOptions = useMemo(
+  const value = useMemo<ProductOptionsHookValue>(
     () => ({
       variants,
-      variantsConnection,
+      variantsConnection: product.variants,
       options,
       selectedVariant,
       setSelectedVariant,
@@ -146,14 +161,14 @@ export function ProductOptionsProvider({
       selectedSellingPlan,
       setSelectedSellingPlan,
       selectedSellingPlanAllocation,
-      // @ts-ignore The types here are broken on main, need to come back and fix them sometime
       sellingPlanGroups,
-      // @ts-ignore The types here are broken on main, need to come back and fix them sometime
-      sellingPlanGroupsConnection,
+      sellingPlanGroupsConnection: product.sellingPlanGroups,
     }),
     [
       isOptionInStock,
       options,
+      product.sellingPlanGroups,
+      product.variants,
       selectedOptions,
       selectedSellingPlan,
       selectedSellingPlanAllocation,
@@ -165,7 +180,7 @@ export function ProductOptionsProvider({
   );
 
   return (
-    <ProductOptionsContext.Provider value={productOptions}>
+    <ProductOptionsContext.Provider value={value}>
       {children}
     </ProductOptionsContext.Provider>
   );
@@ -176,7 +191,7 @@ function getVariantBasedOnIdProp(
   variants: Array<PartialDeep<ProductVariantType> | undefined>
 ) {
   // get the initial variant based on the logic outlined in the comments for 'initialVariantId' above
-  // 1.
+  // * 1. If `initialVariantId` is provided, then it's used even if it's out of stock.
   if (explicitVariantId) {
     const foundVariant = variants.find(
       (variant) => variant?.id === explicitVariantId
@@ -188,12 +203,23 @@ function getVariantBasedOnIdProp(
     }
     return foundVariant;
   }
-  // 2.
+  // * 2. If `initialVariantId` is provided but is `null`, then no variant is used.
   if (explicitVariantId === null) {
     return null;
   }
-  // 3. and 4.
+  // * 3. If nothing is passed to `initialVariantId` then the first available / in-stock variant is used.
+  // * 4. If nothing is passed to `initialVariantId` and no variants are in stock, then the first variant is used.
   if (explicitVariantId === undefined) {
     return variants.find((variant) => variant?.availableForSale) || variants[0];
   }
+}
+
+function isProductVariantArray(
+  maybeVariantArray: (PartialDeep<ProductVariantType> | undefined)[] | undefined
+): maybeVariantArray is PartialDeep<ProductVariantType>[] {
+  if (!maybeVariantArray || !Array.isArray(maybeVariantArray)) {
+    return false;
+  }
+
+  return true;
 }
