+++ packages/hydrogen/src/components/ProductOptionsProvider/tests/fixtures.ts
