+++ templates/demo-store/src/components/ProductDetails.client.jsx
@@ -1,11 +1,10 @@
 import {
-  flattenConnection,
-  useProduct,
+  useProductOptions,
   useParsedMetafields,
-  ProductProvider,
   ProductPrice,
   AddToCartButton,
   BuyNowButton,
+  ProductOptionsProvider,
 } from '@shopify/hydrogen';
 import ProductOptions from './ProductOptions.client';
 import Gallery from './Gallery.client';
@@ -15,7 +14,7 @@ import {
 } from './Button.client';
 
 function AddToCartMarkup() {
-  const {selectedVariant} = useProduct();
+  const {selectedVariant} = useProductOptions();
   const isOutOfStock = !selectedVariant.availableForSale;
 
   return (
@@ -88,26 +87,28 @@ function SizeChart() {
   );
 }
 
-function ProductPrices() {
-  const product = useProduct();
+function ProductPrices({product}) {
+  const {selectedVariant} = useProductOptions();
 
   return (
     <>
       <ProductPrice
         className="text-gray-500 line-through text-lg font-semibold"
         priceType="compareAt"
-        variantId={product.selectedVariant.id}
+        variantId={selectedVariant.id}
+        data={product}
       />
       <ProductPrice
         className="text-gray-900 text-lg font-semibold"
-        variantId={product.selectedVariant.id}
+        variantId={selectedVariant.id}
+        data={product}
       />
     </>
   );
 }
 
 export default function ProductDetails({product}) {
-  const initialVariant = flattenConnection(product.variants)[0];
+  const initialVariant = product.variants.nodes[0];
 
   const productMetafields = useParsedMetafields(product.metafields || {});
   const sizeChartMetafield = productMetafields.find(
@@ -126,7 +127,10 @@ export default function ProductDetails({product}) {
 
   return (
     <>
-      <ProductProvider data={product} initialVariantId={initialVariant.id}>
+      <ProductOptionsProvider
+        data={product}
+        initialVariantId={initialVariant.id}
+      >
         <div className="grid grid-cols-1 md:grid-cols-[2fr,1fr] gap-x-8 my-16">
           <div className="md:hidden mt-5 mb-8">
             <h1 className="text-4xl font-bold text-black mb-4">
@@ -139,11 +143,11 @@ export default function ProductDetails({product}) {
             )}
             <span />
             <div className="flex justify-between md:block">
-              <ProductPrices />
+              <ProductPrices product={product} />
             </div>
           </div>
 
-          <Gallery />
+          <Gallery product={product} />
 
           <div>
             <div className="hidden md:block">
@@ -155,7 +159,7 @@ export default function ProductDetails({product}) {
                   {product.vendor}
                 </div>
               )}
-              <ProductPrices />
+              <ProductPrices product={product} />
             </div>
             {/* Product Options */}
             <div className="mt-8">
@@ -226,7 +230,7 @@ export default function ProductDetails({product}) {
             )}
           </div>
         </div>
-      </ProductProvider>
+      </ProductOptionsProvider>
     </>
   );
 }
