+++ templates/demo-store/src/components/ProductOptions.client.jsx
@@ -1,10 +1,10 @@
-import {useProduct} from '@shopify/hydrogen';
+import {useProductOptions} from '@shopify/hydrogen';
 
 /**
  * A client component that tracks a selected variant and/or selling plan state, as well as callbacks for modifying the state
  */
 export default function ProductOptions() {
-  const {options, setSelectedOption, selectedOptions} = useProduct();
+  const {options, setSelectedOption, selectedOptions} = useProductOptions();
 
   return (
     <>
