+++ templates/demo-store/src/routes/index.server.jsx
@@ -204,47 +204,41 @@ const QUERY = gql`
   query indexContent($country: CountryCode, $language: LanguageCode)
   @inContext(country: $country, language: $language) {
     collections(first: 2) {
-      edges {
-        node {
-          handle
+      nodes {
+        handle
+        id
+        title
+        image {
           id
-          title
-          image {
+          url
+          altText
+          width
+          height
+        }
+        products(first: 3) {
+          nodes {
+            handle
             id
-            url
-            altText
-            width
-            height
-          }
-          products(first: 3) {
-            edges {
-              node {
-                handle
+            title
+            variants(first: 1) {
+              nodes {
                 id
                 title
-                variants(first: 1) {
-                  edges {
-                    node {
-                      id
-                      title
-                      availableForSale
-                      image {
-                        id
-                        url
-                        altText
-                        width
-                        height
-                      }
-                      priceV2 {
-                        currencyCode
-                        amount
-                      }
-                      compareAtPriceV2 {
-                        currencyCode
-                        amount
-                      }
-                    }
-                  }
+                availableForSale
+                image {
+                  id
+                  url
+                  altText
+                  width
+                  height
+                }
+                priceV2 {
+                  currencyCode
+                  amount
+                }
+                compareAtPriceV2 {
+                  currencyCode
+                  amount
                 }
               }
             }
